Fixed builder priority sites, fixed hpPercent

+++ src/main/kotlin/ext/structure.kt
@@ -11,7 +11,7 @@ val Structure.hpDelta get() = hitsMax - hits
 /**
  * The structures percentage of from full hp
  */
-val Structure.hpPercent get() = ((hitsMax - hpDelta) / hitsMax) * 100
+val Structure.hpPercent get() = ((hitsMax - hpDelta).toFloat() / hitsMax.toFloat()) * 100
 
 fun Structure.isFullHp() = hits >= hitsMax
 fun Structure.isMissingHp() = !isFullHp()
